move components to src/pages + fixes
