+++ src/pages/PointOfSale/POS.vue
@@ -69,7 +69,7 @@
 import Transaction from "./Transaction";
 import ItemList from "./ItemList";
 import frappe from "frappejs";
-import FrappeControl from '../controls/FrappeControl';
+import FrappeControl from '../../components/controls/FrappeControl';
 
 export default {
   components: {
